% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic-analyse.R
\name{ptsm_logistic_fitted_delta}
\alias{ptsm_logistic_fitted_delta}
\title{Calculate mean and quantiles of fitted delta values}
\usage{
ptsm_logistic_fitted_delta(sampler_results, probs = c(0.025, 0.1, 0.5, 0.9,
  0.975))
}
\description{
Calculate mean and quantiles of fitted delta values
}
