% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic-analyse.R
\name{ptsm_logistic_melt_delta_samples}
\alias{ptsm_logistic_melt_delta_samples}
\title{Take a wide format theta_samples and perform a melt-like operation on it to split it into a delta for each level}
\usage{
ptsm_logistic_melt_delta_samples(theta_samples, levels, distributions,
  level_var = "level")
}
\description{
Take a wide format theta_samples and perform a melt-like operation on it to split it into a delta for each level
}

